rename to manageContentActions
