+++ kolibri/plugins/management/assets/src/state/manageContentActions.js
